Added views folder and moved UI code out of main
