+++ dartboard/Hub.py
@@ -1,12 +1,17 @@
-from ScorerWindow import ScorerWindow
-from ScoreboardWindow import ScoreboardWindow
-from StartMenuWindow import StartMenuWindow
-from StartGameConfigurationWindow import StartGameConfigurationWindow
-from PySide2.QtWidgets import QApplication, QMainWindow
+from views.ScorerWindow import ScorerWindow
+from views.StartMenuWindow import StartMenuWindow
+from views.StartGameConfigurationWindow import StartGameConfigurationWindow
+from views.ScoreboardWindow import ScoreboardWindow
+
+from PySide2.QtGui import QPalette, QColor
+from PySide2.QtCore import Qt
+from PySide2.QtWidgets import QApplication
 import sys
 
-class Hub():
+class Hub(QApplication):
     def __init__(self):
+        super().__init__()
+        self.set_style()
 
 
         self.start_menu = StartMenuWindow(self)
@@ -22,6 +27,23 @@ class Hub():
         #self.scoreboard.show()
 
     #add functions and high level things here
+    def set_style(self):
+        self.setStyle("Fusion")
+        palette = QPalette()
+        palette.setColor(QPalette.Window, QColor(53, 53, 53))
+        palette.setColor(QPalette.WindowText, Qt.white)
+        palette.setColor(QPalette.Base, QColor(25, 25, 25))
+        palette.setColor(QPalette.AlternateBase, QColor(53, 53, 53))
+        palette.setColor(QPalette.ToolTipBase, Qt.black)
+        palette.setColor(QPalette.ToolTipText, Qt.white)
+        palette.setColor(QPalette.Text, Qt.white)
+        palette.setColor(QPalette.Button, QColor(53, 53, 53))
+        palette.setColor(QPalette.ButtonText, Qt.white)
+        palette.setColor(QPalette.BrightText, Qt.red)
+        palette.setColor(QPalette.Link, QColor(42, 130, 218))
+        palette.setColor(QPalette.Highlight, QColor(42, 130, 218))
+        palette.setColor(QPalette.HighlightedText, Qt.black)
+        self.setPalette(palette)
 
     def navigate_to_view(self, name):
         if (name == "start_menu"):
