+++ dartboard/views/StartGameConfiguration.py
